+++ src/test/java/org/sonar/plugins/delphi/pmd/rules/UnusedImportsRuleTest.java
@@ -120,7 +120,90 @@ public class UnusedImportsRuleTest extends BasePmdRuleTest {
   }
 
   @Test
-  public void testExcludedUnusedImportShouldNotAddIssue() {
+  public void testComponentAncestorImportShouldNotAddIssue() {
+    DelphiTestUnitBuilder builder =
+        new DelphiTestUnitBuilder()
+            .appendDecl("uses")
+            .appendDecl("    Vcl.Controls")
+            .appendDecl("  , System.Classes")
+            .appendDecl("  ;")
+            .appendDecl("type")
+            .appendDecl("  TFooControl = class(TCustomControl)")
+            .appendDecl("  end;")
+            .appendDecl("  TFooComponent = class(TCustomControl)")
+            .appendDecl("    FControl: TFooControl;")
+            .appendDecl("  end;");
+
+    execute(builder);
+
+    assertIssues().areNot(ruleKey("UnusedImportsRule_TEST"));
+  }
+
+  @Test
+  public void testComponentAncestorImportWithPublishedFieldInNonComponentTypeShouldAddIssue() {
+    DelphiTestUnitBuilder builder =
+        new DelphiTestUnitBuilder()
+            .appendDecl("uses")
+            .appendDecl("    Vcl.Controls")
+            .appendDecl("  , System.Classes")
+            .appendDecl("  ;")
+            .appendDecl("type")
+            .appendDecl("  TFooControl = class(TCustomControl)")
+            .appendDecl("  end;")
+            .appendDecl("  TFoo = class(TObject)")
+            .appendDecl("    FControl: TFooControl;")
+            .appendDecl("  end;");
+
+    execute(builder);
+
+    assertIssues()
+        .areExactly(1, ruleKeyAtLine("UnusedImportsRule_TEST", builder.getOffsetDecl() + 3));
+  }
+
+  @Test
+  public void testComponentAncestorImportWithNonPublishedFieldShouldAddIssue() {
+    DelphiTestUnitBuilder builder =
+        new DelphiTestUnitBuilder()
+            .appendDecl("uses")
+            .appendDecl("    Vcl.Controls")
+            .appendDecl("  , System.Classes")
+            .appendDecl("  ;")
+            .appendDecl("type")
+            .appendDecl("  TFooControl = class(TCustomControl)")
+            .appendDecl("  end;")
+            .appendDecl("  TFooComponent = class(TCustomControl)")
+            .appendDecl("  public")
+            .appendDecl("    FControl: TFooControl;")
+            .appendDecl("  end;");
+
+    execute(builder);
+
+    assertIssues()
+        .areExactly(1, ruleKeyAtLine("UnusedImportsRule_TEST", builder.getOffsetDecl() + 3));
+  }
+
+  @Test
+  public void testExcludedUnusedImportInInterfaceSectionShouldNotAddIssue() {
+    property.setValue("System.SysUtils");
+
+    DelphiTestUnitBuilder builder =
+        new DelphiTestUnitBuilder()
+            .appendDecl("uses")
+            .appendDecl("  System.SysUtils;")
+            .appendImpl("procedure Test;")
+            .appendImpl("var")
+            .appendImpl("  Obj: TObject;")
+            .appendImpl("begin")
+            .appendImpl("  Obj := TObject.Create;")
+            .appendImpl("end;");
+
+    execute(builder);
+
+    assertIssues().areNot(ruleKey("UnusedImportsRule_TEST"));
+  }
+
+  @Test
+  public void testExcludedUnusedImportInImplementationSectionShouldAddIssue() {
     property.setValue("System.SysUtils");
 
     DelphiTestUnitBuilder builder =
@@ -132,6 +215,24 @@ public class UnusedImportsRuleTest extends BasePmdRuleTest {
             .appendImpl("  Obj: TObject;")
             .appendImpl("begin")
             .appendImpl("  Obj := TObject.Create;")
+            .appendImpl("end;");
+
+    execute(builder);
+
+    assertIssues().areExactly(1, ruleKeyAtLine("UnusedImportsRule_TEST", builder.getOffset() + 2));
+  }
+
+  @Test
+  public void testImportsRequiredByInlineMethodsShouldNotAddIssue() {
+    DelphiTestUnitBuilder builder =
+        new DelphiTestUnitBuilder()
+            .appendImpl("uses")
+            .appendImpl("    System.UITypes")
+            .appendImpl("  , Vcl.Dialogs")
+            .appendImpl("  ;")
+            .appendImpl("procedure Test;")
+            .appendImpl("begin")
+            .appendImpl("  MessageDlg('Spooky error!', mtError, [mbOK], 0);")
             .appendImpl("end;");
 
     execute(builder);