+++ src/test/resources/org/sonar/plugins/delphi/standardLibrary/Vcl.Controls.pas
@@ -1,0 +1,23 @@
+unit Vcl.Controls;
+
+interface
+
+uses
+  System.Classes;
+
+type
+  TControl = class(TComponent)
+  
+  end;
+  
+  TWinControl = class(TControl)
+  
+  end;
+
+  TCustomControl = class(TWinControl)
+
+  end;
+
+implementation
+
+end.